Dev aof

+++ Ley22_WebApp_V2/Old_App_Code/ControldePago.cs
@@ -25,7 +25,7 @@ namespace Ley22_WebApp_V2.Old_App_Code
         public decimal Cantidad { get; set; }
         public string Descripcion { get; set; }
         public string NumeroRecibo { get; set; }
-        public Nullable<int> NumerodeCheque { get; set; }
+        public Nullable<long> NumerodeCheque { get; set; }
         public string VoidDesc { get; set; }
     }
 }
